// libs/admin-plugins/shipping/domestic-shipping/src/generated-admin-types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
  Date: any;
};
// ...
export type QueryPincodesArgs = {
  options?: Maybe<PincodeListOptions>;
};


export type QueryPincodeArgs = {
  id: Scalars['ID'];
};


export type QueryCheckPincodeArgs = {
  input: CheckPincodeInput;
};
// ...
export type MutationCreatePincodeArgs = {
  input: CreatePincodeInput;
};


export type MutationUpdatePincodeArgs = {
  input: UpdatePincodeInput;
};
// ...
export type PincodeListOptions = {
  skip: Scalars['Int'];
  take: Scalars['Int'];
  sort?: Maybe<PincodeSortParameter>;
  filter?: Maybe<PincodeFilterParameter>;
};

export type CheckPincodeInput = {
  pincode: Scalars['Int'];
  productWeight?: Maybe<Scalars['Int']>;
};

export type UpdatePincodeInput = {
  id: Scalars['ID'];
  pincode?: Maybe<Scalars['Int']>;
  state?: Maybe<Scalars['String']>;
  district?: Maybe<Scalars['String']>;
  prepaid?: Maybe<Scalars['Boolean']>;
  cod?: Maybe<Scalars['Boolean']>;
  pickup?: Maybe<Scalars['Boolean']>;
  cash?: Maybe<Scalars['Boolean']>;
  repl?: Maybe<Scalars['Boolean']>;
};

export type CreatePincodeInput = {
  pincode?: Maybe<Scalars['Int']>;
  state?: Maybe<Scalars['String']>;
  district?: Maybe<Scalars['String']>;
  prepaid?: Maybe<Scalars['Boolean']>;
  cod?: Maybe<Scalars['Boolean']>;
  pickup?: Maybe<Scalars['Boolean']>;
  cash?: Maybe<Scalars['Boolean']>;
  repl?: Maybe<Scalars['Boolean']>;
};
// ...
export type PincodeFilterParameter = {
  createdAt?: Maybe<DateOperators>;
  updatedAt?: Maybe<DateOperators>;
  pincode?: Maybe<NumberOperators>;
  state?: Maybe<StringOperators>;
  district?: Maybe<StringOperators>;
  prepaid?: Maybe<BooleanOperators>;
  cod?: Maybe<BooleanOperators>;
  pickup?: Maybe<BooleanOperators>;
  cash?: Maybe<BooleanOperators>;
  repl?: Maybe<BooleanOperators>;
};

export type PincodeSortParameter = {
  createdAt?: Maybe<SortOrder>;
  updatedAt?: Maybe<SortOrder>;
  id?: Maybe<SortOrder>;
  pincode?: Maybe<SortOrder>;
  state?: Maybe<SortOrder>;
  district?: Maybe<SortOrder>;
};